Extract Redux connect bindings
