+++ src/containers/DatabaseOverview/IngredientsFormPage.tsx
@@ -12,7 +12,6 @@ import {
     TableRow
 } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import { connect } from 'react-redux';
 import SimpleBar from "simplebar-react";
 
 import { Malts } from '../../model/Malt';
@@ -21,14 +20,10 @@ import { Yeasts } from '../../model/Yeasts';
 
 import './IngredientsFormPage.css';
 
-import { MaltsActions } from "../../actions/malt.actions";
-import { HopsActions } from "../../actions/hops.actions";
-import { YeastActions } from "../../actions/yeast.actions";
-import { AdditionalIngredientsActions } from "../../actions/additionalIngredients.actions";
 import { AdditionalIngredient } from "../../model/AdditionalIngredient";
 
 
-class IngredientsFormPage extends React.Component<any, any> {
+export class IngredientsFormPage extends React.Component<any, any> {
 
     simpleBarRef: any = null;
 
@@ -382,27 +377,3 @@ class IngredientsFormPage extends React.Component<any, any> {
 }
 
 /* ====================== Redux Mapper ====================== */
-
-const mapStateToProps = (state: any) => ({
-    malts: state.maltsReducer.malts || [],
-    hops: state.hopsReducer.hops || [],
-    yeasts: state.yeastReducer.yeasts || [],
-    additionalIngredients: state.additionalIngredientsReducer.additionalIngredients || []
-});
-
-const mapDispatchToProps = (dispatch: any) => ({
-    getMalt: (isFetching: boolean) => dispatch(MaltsActions.getMalts(isFetching)),
-    getHop: (isFetching: boolean) => dispatch(HopsActions.getHops(isFetching)),
-    getYeast: (isFetching: boolean) => dispatch(YeastActions.getYeasts(isFetching)),
-    submitNewMalt: (malt: Malts) => dispatch(MaltsActions.submitNewMalt(malt)),
-    submitNewHop: (hop: Hops) => dispatch(HopsActions.submitNewHop(hop)),
-    submitNewYeast: (yeast: Yeasts) => dispatch(YeastActions.submitNewYeast(yeast)),
-    deleteMaltById: (aId: string) => dispatch(MaltsActions.deleteMaltsById(aId)),
-    deleteHopById: (aId: string) => dispatch(HopsActions.deleteHopById(aId)),
-    deleteYeastById: (aId: string) => dispatch(YeastActions.deleteYeastById(aId)),
-    getAdditionalIngredients: (isFetching: boolean) => dispatch(AdditionalIngredientsActions.getAdditionalIngredients(isFetching)),
-    submitNewAdditionalIngredient: (aIngredient: { name: string; description?: string }) => dispatch(AdditionalIngredientsActions.submitNewAdditionalIngredient(aIngredient)),
-    deleteAdditionalIngredientById: (aId: string) => dispatch(AdditionalIngredientsActions.deleteAdditionalIngredientById(aId))
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(IngredientsFormPage);
