+++ src/containers/DatabaseOverview/content/CreateHopForm.tsx
@@ -1,9 +1,6 @@
 import * as React from 'react';
 import {FormEvent} from "react";
 import {Hops} from "../../../model/Hops";
-import {BeerActions} from "../../../actions/actions";
-import {connect} from "react-redux";
-import {HopsActions} from "../../../actions/hops.actions";
 interface HopFormState {
   name: string;
   type: string;
@@ -14,7 +11,7 @@ interface HopFormState {
 interface HopFormProps {
     onSubmitHop: (hop: Hops) => void;
 }
-class HopForm extends React.Component<HopFormProps,HopFormState>
+export class HopForm extends React.Component<HopFormProps,HopFormState>
 {
 constructor(props: HopFormProps) {
     super(props);
@@ -74,12 +71,3 @@ render() {
   }
 
 }
-const mapStateToProps = (state: any) => ({
-    isSuccessful: state.hopsReducer.isSubmitHopSuccessful
-});
-
-const mapDispatchToProps = (dispatch: any) => ({
-    onSubmitHop: (hop: Hops) => dispatch(HopsActions.submitNewHop(hop))
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(HopForm);
