+++ src/containers/DatabaseOverview/content/CreateMaltForm.tsx
@@ -1,10 +1,7 @@
 import * as React from 'react';
 import {ChangeEvent, FormEvent} from "react";
-import {BeerActions} from "../../../actions/actions";
-import {connect} from "react-redux";
 import {Malts} from "../../../model/Malt";
 import {isEqual} from "lodash";
-import {MaltsActions} from "../../../actions/malt.actions";
 
 interface MaltFormState {
     id: string;
@@ -18,7 +15,7 @@ interface MaltFormProps {
     isSuccessful: boolean;
 }
 
-class MaltForm extends React.Component<MaltFormProps,MaltFormState> {
+export class MaltForm extends React.Component<MaltFormProps,MaltFormState> {
     constructor(props: MaltFormProps) {
         super(props);
         this.state = {
@@ -110,13 +107,3 @@ class MaltForm extends React.Component<MaltFormProps,MaltFormState> {
 
 
 }
-
-
-const mapStateToProps = (state: any) => ({
-    isSuccessful: state.beerDataReducer.isSubmitMaltSuccessful
-});
-const mapDispatchToProps = (dispatch: any) => ({
-    onSubmitMalt: (malt: Malts) => dispatch(MaltsActions.submitNewMalt(malt)),
-
-});
-export default connect(mapStateToProps,mapDispatchToProps)(MaltForm);
