+++ src/containers/MainView/Main.connect.ts
@@ -1,0 +1,14 @@
+import {connect} from "react-redux";
+import {BeerActions} from "../../actions/actions";
+import getBeers = BeerActions.getBeers;
+import {Main} from './Main';
+
+const mapStateToProps = (state: any) => ({beers: state.beerDataReducer.beers});
+const mapDispatchToProps =
+    (dispatch: any) => ({
+        getBeers: (isFetching: boolean) => dispatch(getBeers(isFetching))
+    });
+
+
+
+export default connect(mapStateToProps,mapDispatchToProps)(Main);
